create xclim.indicators.convert and xclim.indices.converters, allow for backwards-compatibility, assign corn_heat_units as a converter, load xclim.indcators.convert on init
