+++ src/xclim/__init__.py
@@ -9,7 +9,7 @@ from xclim.core import units  # noqa
 from xclim.core.indicator import build_indicator_module_from_yaml
 from xclim.core.locales import load_locale
 from xclim.core.options import set_options  # noqa
-from xclim.indicators import atmos, generic, land, seaIce  # noqa
+from xclim.indicators import atmos, convert, generic, land, seaIce  # noqa
 
 __author__ = """Travis Logan"""
 __email__ = "logan.travis@ouranos.ca"
