+++ src/xclim/indicators/convert/__init__.py
@@ -1,0 +1,13 @@
+"""
+Conversion Indicators
+=====================
+
+This submodule contains indicators that convert units of input data to other units.
+For example, converting temperature from Kelvin to Celsius or precipitation from kg m-2 s-1 to mm day-1.
+It also includes indicators that convert between related physical quantities, such as converting
+snow depth to snow water equivalent or wind speed to wind power density.
+"""
+
+from __future__ import annotations
+
+from ._conversion import *
