+++ src/xclim/testing/helpers.py
@@ -15,7 +15,7 @@ from dask.callbacks import Callback
 import xclim
 from xclim.core import VARIABLES
 from xclim.core.calendar import percentile_doy
-from xclim.indices import (
+from xclim.indices.converters import (
     longwave_upwelling_radiation_from_net_downwelling,
     shortwave_upwelling_radiation_from_net_downwelling,
 )
